Remove timezone from timestamps in test metadata reference files
The tests currently only work in the timezone of the original author
because the metadata reference files contain explicit timezone info,
while extraction of .CAB files treat the timestamps as local.

This change removes the timezone from the timestamps in the test
references, so that the tests can be run successfully anywhere in the
world.