feat: update lessmsi to use .NET Framework 4.8 or newer and be developed with Visual Studio 2022 (#189)
BREAKING CHANGE: Versions of .NET Framework before are no longer supported by lessmsi. If you want to use an older version use the version v1.11.3 of lessmsi or an earlier release. The lessmsi README (https://github.com/activescott/lessmsi) documents the support policy going forward, which is basically to support the oldest version of Windows that Microsoft officially supports (currently Windows 10) and a version of .NET Framework that is included with that oldest supported version of Windows.